* 수정사항 1. 관광명소 기능    - 리스트 조회, 개별 조회, 개별 등록/수정, 삭제기능, 이미지 업로드, 이미지 삭제 2. 시스템 관리자는 회원탈퇴 불가능하도록 수정

+++ controller/system.js
@@ -70,9 +70,16 @@ exports.deleteUser = async (req, res, next) => {
     const regUserGuid = helper.getsessionValueOrRequsetValue(req.session.userGuid, req.body.userGuid);
     
     try {
+
+      //사용자 계정이 시스템 관리자인 경우
+      if(req.session.authGroupCode == 'S'){
+        resModel = helper.createResponseModel(false, '시스템 관리자는 삭제할 수 없습니다.', '');
+        return res.status(200).json(resModel);
+      }
+
       //오늘의 출장 삭제
       let retVal = await userService.deleteUserList(userGuidList, 'Y', regUserGuid);
-  
+
       //삭제
       if (retVal == 1) {
         req.session.destroy(function(){
